AI_Agency/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
aiofiles==24.1.0
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
